+++ bricks/backend_setup_with_jwt/__brick__/lib/setup/setup.dart
@@ -1,10 +1,10 @@
 import 'package:dio/dio.dart';
 import 'package:get_it/get_it.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:{{project_name}}/api_utils/interceptors/custom_dio_interceptor.dart';
-import 'package:{{project_name}}/services/auth_service.dart';
-import 'package:{{project_name}}/services/platform_services/storage_services/token_storage_service.dart';
-import 'package:{{project_name}}/routes/routes.dart';
+import 'package:{{project-name}}/api_utils/interceptors/custom_dio_interceptor.dart';
+import 'package:{{project-name}}/services/auth_service.dart';
+import 'package:{{project-name}}/services/platform_services/storage_services/token_storage_service.dart';
+import 'package:{{project-name}}/routes/routes.dart';
 
 final getIt = GetIt.instance;