+++ bricks/basic_setup/__brick__/lib/home/home_screen.dart
@@ -1,6 +1,6 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
-import 'package:{{project_name}}/l10n/l10n.dart';
+import 'package:{{project-name}}/l10n/l10n.dart';
 
 @RoutePage()
 class HomeScreen extends StatelessWidget {