+++ bricks/basic_setup/__brick__/lib/routes/routes.dart
@@ -1,6 +1,6 @@
 import 'package:auto_route/annotations.dart';
 import 'package:auto_route/auto_route.dart';
-import 'package:{{project_name}}/home/home_screen.dart';
+import 'package:{{project-name}}/home/home_screen.dart';
 part 'routes.gr.dart';
 
 @AutoRouterConfig(replaceInRouteName: 'Screen,Route')