+++ bricks/basic_setup/__brick__/lib/setup/setup.dart
@@ -1,5 +1,5 @@
 import 'package:get_it/get_it.dart';
-import 'package:{{project_name}}/routes/routes.dart';
+import 'package:{{project-name}}/routes/routes.dart';
 
 final getIt = GetIt.instance;