+++ lib/src/commands/create/create_command.dart
@@ -29,21 +29,22 @@ class CreateCommand extends Command<int> {
         abbr: 'i',
         help: 'Application name',
       )
-      ..addOption(
-        'firebase-project-id',
-        abbr: 'f',
-        help: 'The ID of an existing Firebase project',
-      )
-      ..addFlag('has-external-backend', abbr: 'e', negatable: false)
       ..addFlag(
-        'use-bloc',
-        abbr: 'b',
-        help: 'Implement bloc state management',
+        'integrate-firebase',
+        abbr: 'f',
+        help: 'Integrate a firebase project',
+        negatable: false,
+      )
+      ..addFlag(
+        'has-external-backend',
+        abbr: 'e',
+        negatable: false,
       )
       ..addFlag(
         'use-riverpod',
         abbr: 'r',
         help: 'Implement Riverpod state management',
+        negatable: false,
       )
       ..addFlag(
         'use-deep-linking',
@@ -82,11 +83,14 @@ class CreateCommand extends Command<int> {
 
         late String newDirectoryPath;
 
-        // If the last path segment is empty, use the second-to-last path segment as the new directory name
+        // If the last path segment is empty, use the second-to-last path
+        // segment as the new directory name
         if ((lastPath == null || lastPath.trim().isEmpty) &&
             pathSegments.length > 1) {
           newDirectoryPath =
-              '${destination.path}/${entity.uri.pathSegments[entity.uri.pathSegments.length - 2]}';
+              '${destination.path}/${entity.uri.pathSegments[
+                entity.uri.pathSegments.length - 2
+              ]}';
         } else {
           newDirectoryPath = '${destination.path}/${pathSegments.last}';
         }
@@ -100,18 +104,75 @@ class CreateCommand extends Command<int> {
     }
   }
 
+  Future<bool> deleteFolder(String folderPath) async {
+    final directory = Directory(folderPath);
+
+    if (directory.existsSync()) {
+      try {
+        await directory.delete(recursive: true);
+        _logger.success('Successfully deleted folder: $folderPath');
+        return true;
+      } catch (e) {
+        _logger.err('Error deleting folder: $e');
+        return false;
+      }
+    } else {
+      _logger.err('Folder does not exist: $folderPath');
+      return false;
+    }
+  }
+
   @override
   Future<int> run() async {
     String? projectName = argResults?['project-name'] as String?;
     final applicationId =
         (argResults?['application-id'] as String?) ?? 'com.example.app';
     final applicationName = argResults?['application-name'] as String?;
-    final firebaseProjectId = argResults?['firebase-project-id'] as String?;
+    final integrateFirebase = argResults?['integrate-firebase'] as bool;
     final useRiverpod = argResults?['use-riverpod'] as bool;
-    final useBloc = argResults?['use-bloc'] as bool;
     final hasExternalBackend = argResults?['has-external-backend'] as bool;
     final addPushNotifications = argResults?['add-push-notifications'] as bool;
     final useDeepLinking = argResults?['use-deep-linking'] as bool;
+
+    // Ensure the user has firebase installed if the firebase option is selected
+    if (integrateFirebase) {
+      // Run the 'firebase --version' command
+      final firebaseResult = await Process.run('firebase', ['--version']);
+
+      // If the exit code is not 0, the command failed
+      if (firebaseResult.exitCode != 0) {
+        usageException(
+          'Firebase CLI is not installed or not in PATH. You can '
+          'install it here: https://firebase.google.com/docs/cli',
+        );
+      }
+
+      // check if the user has the flutterfire CLI tool installed
+      final flutterfireResult = await Process.run('flutterfire', ['--version']);
+
+      // If the exit code is not 0, the command failed
+      if (flutterfireResult.exitCode != 0) {
+        // Install the flutterfire CLI tool
+        _logger.info(blue.wrap('Installing flutterfire CLI tool...'));
+        final installFlutterFire = await Process.run(
+          'dart',
+          [
+            'pub',
+            'global',
+            'activate',
+            'flutterfire_cli',
+          ],
+        );
+
+        // check if the command failed
+        if (installFlutterFire.exitCode != 0) {
+          _logger.err('Failed to install flutterfire CLI tool.');
+          return ExitCode.software.code;
+        }
+
+        _logger.success('flutterfire CLI tool installed.');
+      }
+    }
 
     // Check if the project name was provided
     if (projectName == null) {
@@ -126,18 +187,14 @@ class CreateCommand extends Command<int> {
         }
 
         _logger.err(
-          'Invalid project name. Please use only lowercase letters and underscores.',
+          'Invalid project name. Please use only lowercase letters and '
+              'underscores.',
         );
       }
     }
 
-    // Ensure that only one of the Riverpod or Bloc options is selected
-    if (useRiverpod && useBloc) {
-      usageException(
-        'You can only specify one of the Riverpod or Bloc options',
-      );
-    }
-
+    // Check if the user wants to implement deep linking and get
+    // the deeplink URI
     Uri? deeplinkUri;
 
     if (useDeepLinking) {
@@ -146,7 +203,10 @@ class CreateCommand extends Command<int> {
         stdout.write(
           'Enter your deeplink url (https://example.com): ',
         );
+        // read the input from the user
         final deeplinkUriInput = stdin.readLineSync();
+
+        // check if the input is a valid URI
         if (deeplinkUriInput != null &&
             Uri.tryParse(deeplinkUriInput) != null) {
           deeplinkUri = Uri.parse(deeplinkUriInput);
@@ -154,31 +214,53 @@ class CreateCommand extends Command<int> {
         }
 
         _logger.err(
-          'Invalid deep link. Please ensure it has the format ((https://example.com)).',
+          'Invalid deep link. Please ensure it has the format '
+              '((https://example.com)).',
         );
       }
     }
 
-    // Check if the user has installed the very_good_cli package
-    _logger.info('Creating very_good flutter app...');
-    final createVeryGoodFlutterApp = await Process.start('very_good', [
-      'create',
-      'flutter_app',
-      projectName,
-      '--application-id',
-      applicationId,
-    ], mode: ProcessStartMode.inheritStdio);
+    // Create the very_good flutter_app
+    _logger.info(blue.wrap('Creating very_good flutter app...'));
+    final createVeryGoodFlutterApp = await Process.start(
+      'very_good',
+      [
+        'create',
+        'flutter_app',
+        projectName,
+        '--application-id',
+        applicationId,
+      ],
+      mode: ProcessStartMode.inheritStdio,
+    );
 
+    // Check if the create command failed
     if ((await createVeryGoodFlutterApp.exitCode) != 0) {
-      _logger.err('Failed to create very_good_flutter_app. '
-          '${createVeryGoodFlutterApp.stderr}',);
+      _logger.err(
+        'Failed to create very_good_flutter_app. '
+        '${createVeryGoodFlutterApp.stderr}',
+      );
       return ExitCode.software.code;
     }
-    _logger
-      ..info('very_good_flutter_app have been setup.')
 
-      // Install the required packages for setup
-      ..info('Installing required packages...');
+    _logger
+      ..success('very_good_flutter_app have been setup.')
+
+      // Delete the counter folders
+      ..info(blue.wrap('Deleting counter folders...'));
+    final deleteAppCounterFolder =
+        await deleteFolder('$projectName/lib/counter');
+    final deleteTestCounterViewFolder =
+        await deleteFolder('$projectName/test/counter');
+
+    // check if any of them failed to delete
+    if (!deleteAppCounterFolder || !deleteTestCounterViewFolder) {
+      _logger.err('Failed to delete counter folders.');
+      return ExitCode.software.code;
+    }
+
+    // Install the required packages for setup
+    _logger.info(blue.wrap('Installing required packages...'));
     await Process.run(
       'flutter',
       [
@@ -188,6 +270,7 @@ class CreateCommand extends Command<int> {
         'get_it',
         'auto_route',
         'equatable',
+        useRiverpod ? 'flutter_riverpod' : '',
       ],
       workingDirectory: projectName,
     );
@@ -199,6 +282,7 @@ class CreateCommand extends Command<int> {
         'pub',
         'remove',
         'flutter_gen',
+        ...(useRiverpod ? ['flutter_bloc', 'bloc', 'bloc_test'] : []),
       ],
       workingDirectory: projectName,
     );
@@ -217,12 +301,13 @@ class CreateCommand extends Command<int> {
     );
 
     if (applicationName != null) {
-      _logger.info('Setting application name...');
+      _logger.info(blue.wrap('Setting application name...'));
       //   TODO: Change Application name
     }
 
-    // Create basic setup files which include Get_it, Equatable, and AutoRoute setup
-    _logger.info('Setting up setup files...');
+    // Create basic setup files which include Get_it, Equatable,
+    // and AutoRoute setup
+    _logger.info(blue.wrap('Setting up setup files...'));
 
     final makeBasicSetup = await Process.start(
       'mason',
@@ -231,28 +316,31 @@ class CreateCommand extends Command<int> {
         'basic_setup',
         '-o',
         projectName,
-        '--project_name',
+        '--project-name',
         projectName,
         '--on-conflict',
         'overwrite',
-        '--deep_link_scheme',
+        '--deeplink-scheme',
         deeplinkUri?.scheme ?? '',
-        '--deep_link_host',
+        '--deeplink-host',
         deeplinkUri?.host ?? '',
+        '--use-riverpod',
+        useRiverpod.toString(),
       ],
       mode: ProcessStartMode.inheritStdio,
     );
 
     if ((await makeBasicSetup.exitCode) != 0) {
       _logger.err(
-          'Failed to setup setup files. ${makeBasicSetup.stderr.toString()}');
+        'Failed to setup setup files. ${makeBasicSetup.stderr}',
+      );
       return ExitCode.software.code;
     }
     _logger
-      ..info('Setup files have been setup.')
+      ..success('Setup files have been setup.')
 
       // Execute build runner to build the auto_route files
-      ..info('Building auto_route files...');
+      ..info(blue.wrap('Building auto_route files...'));
     final buildRunner = await Process.run(
       'dart',
       [
@@ -284,9 +372,10 @@ class CreateCommand extends Command<int> {
       workingDirectory: projectName,
     );
 
-    // check if the user opted for an external backend and create the api_utils folder
+    // check if the user opted for an external backend and create the
+    // api_utils folder
     if (hasExternalBackend) {
-      _logger.info('Setting up api_utils...');
+      _logger.info(blue.wrap('Setting up api_utils...'));
 
       final makeSetupWithExternalBackend = await Process.start(
         'mason',
@@ -295,7 +384,7 @@ class CreateCommand extends Command<int> {
           'basic_setup_with_external_backend',
           '-o',
           projectName,
-          '--project_name',
+          '--project-name',
           projectName,
           '--on-conflict',
           'overwrite',
@@ -306,7 +395,7 @@ class CreateCommand extends Command<int> {
       if ((await makeSetupWithExternalBackend.exitCode) != 0) {
         _logger.err(
           'Failed to setup external backend setup files. '
-              '${makeSetupWithExternalBackend.stderr.toString()}',
+          '${makeSetupWithExternalBackend.stderr}',
         );
         return ExitCode.software.code;
       }
@@ -329,9 +418,10 @@ class CreateCommand extends Command<int> {
         }
       }
 
-      // check if the user opted for an external backend with JWT and create the api_utils folder
+      // check if the user opted for an external backend with JWT and create
+      // the api_utils folder
       if (usesJWT == 'y') {
-        _logger.info('Installing flutter_secure_storage...');
+        _logger.info(blue.wrap('Installing flutter_secure_storage...'));
         await Process.run(
           'flutter',
           [
@@ -342,7 +432,7 @@ class CreateCommand extends Command<int> {
           workingDirectory: projectName,
         );
 
-        _logger.info('Setting up API Interceptor...');
+        _logger.info(blue.wrap('Setting up API Interceptor...'));
 
         final makeSetupJWT = await Process.start(
           'mason',
@@ -351,7 +441,7 @@ class CreateCommand extends Command<int> {
             'backend_setup_with_jwt',
             '-o',
             projectName,
-            '--project_name',
+            '--project-name',
             projectName,
             '--on-conflict',
             'overwrite',
@@ -361,17 +451,28 @@ class CreateCommand extends Command<int> {
 
         if ((await makeSetupJWT.exitCode) != 0) {
           _logger.err(
-            'Failed to setup API Interceptor for JWT authentication. ${makeSetupJWT.stderr.toString()}',
+            'Failed to setup API Interceptor for JWT authentication. '
+            '${makeSetupJWT.stderr}',
           );
           return ExitCode.software.code;
         }
-        _logger.info('API Interceptor for JWT authentication has been setup.');
+        _logger.success(
+          'Note: API, Auth, and TokenStorage services as well as the API '
+          'Interceptor have been configured, and you will have to '
+          'implement the refreshToken function in the AuthService.',
+        );
       }
     }
 
     // Check if the user wants to add notifications
     if (addPushNotifications) {
-      _logger.info('Installing firebase_messaging, firebase_core, and flutter_local_notifications...');
+      _logger.info(
+        blue.wrap(
+          'Installing firebase_messaging, firebase_core, '
+          'and flutter_local_notifications...',
+        ),
+      );
+
       await Process.run(
         'flutter',
         [
@@ -384,7 +485,7 @@ class CreateCommand extends Command<int> {
         workingDirectory: projectName,
       );
 
-      _logger.info('Setting up push notifications...');
+      _logger.info(blue.wrap('Setting up push notifications...'));
 
       final makeSetupPushNotifications = await Process.start(
         'mason',
@@ -393,27 +494,89 @@ class CreateCommand extends Command<int> {
           'notification_setup',
           '-o',
           projectName,
-          '--project_name',
+          '--project-name',
           projectName,
-          '--application_id',
+          '--application-id',
           applicationId,
           '--on-conflict',
           'overwrite',
-          '--deep_link_scheme',
+          '--deeplink-scheme',
           deeplinkUri?.scheme ?? '',
-          '--deep_link_host',
+          '--deeplink-host',
           deeplinkUri?.host ?? '',
+          '--use-riverpod',
+          useRiverpod.toString(),
         ],
         mode: ProcessStartMode.inheritStdio,
       );
 
       if ((await makeSetupPushNotifications.exitCode) != 0) {
         _logger.err(
-          'Failed to setup notifications. ${makeSetupPushNotifications.stderr.toString()}',
+          'Failed to setup notifications. ${makeSetupPushNotifications.stderr}',
         );
         return ExitCode.software.code;
       }
-      _logger.info('Notifications have been setup.');
+      _logger.success('Notifications have been setup.');
+    }
+
+    // Run dart fix command to fix any lint issues
+    _logger.info(blue.wrap('Running dart fix --apply...'));
+    await Process.run(
+      'dart',
+      [
+        'fix',
+        '--apply',
+      ],
+      workingDirectory: projectName,
+    );
+
+    // Remove the flutter_gen package created by the dart fix --apply
+    await Process.run(
+      'flutter',
+      [
+        'pub',
+        'remove',
+        'flutter_gen',
+      ],
+      workingDirectory: projectName,
+    );
+
+    // Project setup complete message
+    _logger.success(
+      'Project setup complete! To run your project, '
+      "navigate to the project directory and use 'flutter run'",
+    );
+
+    // State management choice
+    if (useRiverpod) {
+      _logger.success(
+        'Note: Basic Riverpod setup has been done, but you may need to manually'
+            ' adjust some code for full Riverpod integration.',
+      );
+    }
+
+    // Firebase choice
+    if (integrateFirebase) {
+      _logger.success(
+          'Note: Firebase integration has been set up, but you will need to '
+          'manually configure the Firebase project for staging and production'
+          ' environments.');
+    }
+
+    // Messaging choice
+    if (addPushNotifications) {
+      _logger.success(
+        'Note: Firebase Messaging notifications have been set up, but you may'
+        ' need to implement the processNotification method in '
+        'the NotificationService.',
+      );
+    }
+
+    // Dynamic Link choice
+    if (useDeepLinking) {
+      _logger.success('Note: Dynamic Links have been set up, navigate to your '
+          'AndroidManifest.xml and Info.plist to update the links, and to the '
+          'appRouterConfig variable in the app.dart to configure them.');
     }
 
     return ExitCode.success.code;