+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Flutter Project Setup CLI Changelog
 
 - Implemented flutter_project_setup_cli create command template with specified args
 - Implemented dynamic links, and JWT authentication setups
+- Implemented State Management setup with Riverpod